requirements.txt
asgiref==3.5.2
coreapi==2.3.3
Django==4.0.6
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
django-rest-swagger==2.2.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.3
Jinja2==3.1.2
Pillow==9.2.0
psycopg2-binary==2.9.3
requests==2.28.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
gunicorn
whitenoise==6.2.0
MarkupSafe==2.1.1
cloudinary==1.29.0

